+++ CHANGELOG.md
@@ -7,10 +7,11 @@
     - Rename `ActionAcceptedOutput` to `AcceptedActionOutput`, and make it a class, rewrite its logics.
     - `Action.mute/check/reject/then` will take `AcceptedActionOutput` as the callback's param.
     - Regex should not starts with `^` and will auto have sticky flag.
-    - Export `Action.simple/match`. Add `options.autoSticky/rejectCaret` for `Action.match`.
-    - Add `options.autoSticky/autoGlobal` for `fromTo`.
+    - Lexer's constructor is now private.
   - Feat: `SimpleActionExec` can accept `string/SimpleAcceptedActionOutput` as the output.
   - Feat: add generic parameter for Lexer family as the `Token.error` type.
+  - Feat: add `options.autoSticky/autoGlobal` for `fromTo`.
+  - Feat: export `Action.simple/match`. Add `options.autoSticky/rejectCaret` for `Action.match`.
 - Optimize performance & comments.
 
 ## v0.11.0